+++ CHANGELOG.md
@@ -5,7 +5,19 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.6 - Unreleased
+## Version 0.2.7 - Unreleased
+
+### Added
+
+* Added "domain net" to `kwcoco grab`.
+* Added "spacenet7" to `kwcoco grab`.
+
+
+### Fixed
+* Fixed issue in `add_annotation` where if failed to add keypoints, and segmentations.
+
+
+## Version 0.2.6 - Released 2021-06-22
 
 ### Added